requirements.txt
ai2thor==5.0.0
prior
gymnasium==0.29.1
imageio==2.33.1
scikit-learn==1.6.1
imageio-ffmpeg==0.4.9
matplotlib
matplotlib-inline
numpy==1.21.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pandas
pillow
seaborn
torch>=2.2.0
torchvision>=0.17.0
tqdm
gdown
sciex
Cython
sb3_contrib
tensorboard
jupyter
